requirements.txt
fastapi==0.135.3
gunicorn==25.3.0
langchain==1.2.15
langchain-core==1.2.26
langchain-openai==1.1.12
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
openai==2.30.0
pydantic==2.12.5
python-dotenv==1.2.2
tiktoken==0.12.0
typing_extensions==4.15.0
uvicorn==0.43.0
